+++ lib/esm/index.d.ts
@@ -1,7 +1,7 @@
-export { editAt, editWhere, filterBy, removeAt, removeWhere, searchedItems, sortByKey, splitArray, insert, insertAt } from './array';
+export { editAt, editWhere, filterBy, removeAt, removeWhere, searchedItems, sortByKey, splitArray, insert, insertAt, shuffle, groupByKey } from './array';
 export { convertTimestampFromUnixEpoch, formatDateBR, formatTimeBR, getUnixEpochTimestamp } from './date';
 export { maskCep, maskCnpj, maskCpf, maskCreditCard, maskDate, maskMoney, maskNumber, maskNumberDecimals, maskPhone, maskText, maskPassword } from './mask';
-export { keySelect, getValueAtPath, setValueAtPath, jsonToBlob, blobToJson } from './object';
+export { keySelect, getValueAtPath, setValueAtPath, jsonToBlob, blobToJson, remapProperties } from './object';
 export { capitalize, extensionFromFileName, fileNameFromPath, formatMoneyBR, formatPlacaOld, normalizeLower, numberFromText, titleize } from './string';
 export { bytesToSize, getCreditCardBrand } from './util';
 export { cepIsValid, cnpjIsValid, cpfIsValid, cpfcnpjIsValid, emailIsValid, phoneIsValid, uuidIsValid } from './validate';